+++ src/components/BurgerLine/BurgerLine.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import * as S from "./Styles";
+import * as S from "./styles";
 
 export function BurgerLine() {
   return <S.BurgerLine></S.BurgerLine>;
