+++ src/components/BurgerLine/styles.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const BurgerLine = styled.span`
+  display: inline-block;
+  width: 100%;
+  height: 1px;
+  background-color: #d3d3d3;
+`;
