+++ src/components/Container/styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  max-width: 1920px;
+  height: 100vh;
+  margin: 0 auto;
+  position: relative;
+  background-color: #181818;
+`;
