+++ src/components/ContentHeader/ContentHeader.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import * as S from "./Styles";
+import * as S from "./styles";
 
 export const ContentHeader = () => {
   return (
