+++ src/components/MainNav/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const MainNav = styled.nav`
+  width: 244px;
+  background-color: #181818;
+  padding: 20px 0 20px 36px;
+`;
