+++ src/components/NavLogo/styles.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const NavLogo = styled.div`
+  width: 113.33px;
+  height: 43px;
+  padding: 13px 0 13px 0;
+  background-color: transparent;
+  margin-bottom: 20px;
+`;
+
+export const NavImg = styled.img`
+  width: 113.33px;
+  height: 17px;
+  color: #181818;
+`;
