+++ src/components/NavMenu/styles.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const MenuNav = styled.nav`
+  display: block;
+  visibility: visible;
+`;
+
+export const MenuList = styled.ul`
+  padding: 18px 0 10px 0;
+`;
