+++ src/components/NavMenuItem/NavMenuItem.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import * as S from "./Styles"
+import * as S from "./styles";
 
 export const NavMenuItem = ({ text, href }) => {
   return (
