+++ src/components/NavMenuItem/styles.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const MenuItem = styled.li`
+  padding: 5px 0;
+  margin-bottom: 16px;
+`;
+
+export const MenuLink = styled.a`
+  color: #ffffff;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+`;
