--- src/components/Styles.js
@@ -1,1021 +1,0 @@
-import styled from "styled-components";
-
-//Bar
-export const Bar = styled.div`
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  background: rgba(28, 28, 28, 0.5);
-`;
-
-export const BarContent = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-`;
-
-export const BarPlayerProgress = styled.div`
-  width: 100%;
-  height: 5px;
-  background: #2e2e2e;
-`;
-
-export const BarPlayerBlock = styled.div`
-  height: 73px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-`;
-
-export const BarPlayer = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: start;
-  -ms-flex-pack: start;
-  justify-content: flex-start;
-`;
-
-export const PlayerControls = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  padding: 0 27px 0 31px;
-`;
-
-export const PlayerBtnPrev = styled.div`
-  padding: 5px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-right: 23px;
-`;
-
-export const PlayerBtnPrevSvg = styled.svg`
-  width: 15px;
-  height: 14px;
-`;
-
-export const PlayerBtnPlay = styled.div`
-  padding: 5px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-right: 23px;
-  cursor: pointer;
-`;
-
-export const PlayerBtnPlaySvg = styled.svg`
-  width: 22px;
-  height: 20px;
-  fill: #d9d9d9;
-`;
-
-export const PlayerBtnNext = styled.div`
-  padding: 5px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-right: 28px;
-  fill: #a53939;
-`;
-
-export const PlayerBtnNextSvg = styled.svg`
-  width: 15px;
-  height: 14px;
-  fill: inherit;
-  stroke: #d9d9d9;
-`;
-
-export const PlayerBtnRepeat = styled.div`
-  padding: 5px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-right: 24px;
-`;
-
-export const PlayerBtnRepeatSvg = styled.svg`
-  width: 18px;
-  height: 12px;
-  fill: transparent;
-  stroke: #696969;
-`;
-
-export const PlayerBtnShuffle = styled.div`
-  padding: 5px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-`;
-
-export const PlayerBtnShuffleSvg = styled.svg`
-  width: 19px;
-  height: 12px;
-  fill: transparent;
-  stroke: #696969;
-`;
-
-export const BarVolumeBlock = styled.div`
-  width: auto;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  padding: 0 92px 0 0;
-`;
-
-export const VolumeContent = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: end;
-  -ms-flex-pack: end;
-  justify-content: end;
-`;
-
-export const VolumeImg = styled.div`
-  width: 13px;
-  height: 18px;
-  margin-right: 17px;
-`;
-
-export const VolumeSvg = styled.svg`
-  width: 13px;
-  height: 18px;
-  fill: transparent;
-`;
-
-export const VolumeProgress = styled.div`
-  width: 109px;
-`;
-
-export const VolumeProgressLine = styled.input`
-  width: 109px;
-`;
-
-//BarTrackPlay
-export const BarPlayerTrackPlay = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-`;
-
-export const BarTrackPlayContain = styled.div`
-  width: auto;
-  display: -ms-grid;
-  display: grid;
-  -ms-grid-columns: auto 1fr;
-  grid-template-columns: auto 1fr;
-  grid-template-areas: "image author" "image album";
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-`;
-
-export const BarTrackPlayImg = styled.div`
-  width: 51px;
-  height: 51px;
-  background-color: #313131;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
-  margin-right: 12px;
-  -ms-grid-row: 1;
-  -ms-grid-row-span: 2;
-  -ms-grid-column: 1;
-  grid-area: image;
-`;
-
-export const BarTrackPlaySvg = styled.svg`
-  width: 18px;
-  height: 17px;
-  fill: transparent;
-  stroke: #4e4e4e;
-`;
-
-export const BarTrackPlayAuthor = styled.div`
-  -ms-grid-row: 1;
-  -ms-grid-column: 2;
-  grid-area: author;
-  min-width: 49px;
-`;
-
-export const BarTrackPlayAuthorLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-  white-space: nowrap;
-`;
-
-export const BarTrackPlayAlbum = styled.div`
-  -ms-grid-row: 2;
-  -ms-grid-column: 2;
-  grid-area: album;
-  min-width: 49px;
-`;
-
-export const BarTrackPlayAlbumLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 13px;
-  line-height: 24px;
-  color: #ffffff;
-`;
-
-export const BarTrackPlayLikeDis = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-left: 26%;
-`;
-
-export const BarTrackPlayLike = styled.div`
-  padding: 5px;
-`;
-
-export const BarTrackPlayLikeSvg = styled.svg`
-  width: 14px;
-  height: 12px;
-  fill: transparent;
-  stroke: #696969;
-`;
-
-export const BarTrackPlayDislike = styled.div`
-  padding: 5px;
-  margin-left: 28.5px;
-`;
-
-export const BarTrackPlayDislikeSvg = styled.svg`
-  width: 14.34px;
-  height: 13px;
-  fill: transparent;
-  stroke: #696969;
-`;
-
-//BurgerLine
-export const BurgerLine = styled.span`
-  display: inline-block;
-  width: 100%;
-  height: 1px;
-  background-color: #d3d3d3;
-`;
-
-//BurgerNav
-export const NavBurger = styled.div`
-  width: 20px;
-  height: 36px;
-  padding: 13px 0;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-`;
-
-//ContentHeader
-export const ContentTitle = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-  margin-bottom: 24px;
-`;
-
-export const PlaylistTitleCol = styled.div`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 24px;
-  letter-spacing: 2px;
-  color: #696969;
-  text-transform: uppercase;
-`;
-
-export const PlaylistTitleCol1 = styled(PlaylistTitleCol)`
-  width: 447px;
-`;
-
-export const PlaylistTitleCol2 = styled(PlaylistTitleCol)`
-  width: 321px;
-`;
-
-export const PlaylistTitleCol3 = styled(PlaylistTitleCol)`
-  width: 245px;
-`;
-
-export const PlaylistTitleCol4 = styled(PlaylistTitleCol)`
-  width: 60px;
-  text-align: end;
-`;
-
-export const PlaylistTitleSvg = styled.svg`
-  width: 12px;
-  height: 12px;
-  fill: transparent;
-  stroke: #696969;
-`;
-
-//Filter
-export const CenterBlockFilter = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-bottom: 51px;
-`;
-
-export const FilterTitle = styled.div`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  margin-right: 15px;
-`;
-
-//FilterButton
-export const FilterButtonGroup = styled.div`
-  position: relative;
-  &:not(:last-child) {
-    margin-right: 10px;
-  }
-`;
-
-export const FilterButton = styled.button`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  border: 1px solid #ffffff;
-  border-radius: 60px;
-  padding: 6px 20px;
-  background-color: #000000;
-  color: #fff;
-  &:active {
-    color: #ad61ff;
-    border-color: #ad61ff;
-    cursor: pointer;
-  }
-  &:hover {
-    border-color: #d9b6ff;
-    color: #d9b6ff;
-    cursor: pointer;
-    border-radius: 60px;
-    border: 1px solid var(--palette-purple-40, #d9b6ff);
-    cursor: pointer;
-  }
-`;
-
-export const FilterDropdown = styled.div`
-  position: absolute;
-  margin-top: 10px;
-  width: 248px;
-  max-height: 305px;
-  background-color: green;
-  -webkit-box-sizing: border-box;
-  padding: 34px;
-  gap: 10px;
-  border-radius: 12px;
-  background: #313131;
-  overflow-y: scroll;
-`;
-
-export const FilterListElement = styled.div`
-  color: #fff;
-  font-variant-numeric: lining-nums proportional-nums;
-  font-family: StratosSkyeng;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 24px;
-  &:hover {
-    color: #b672ff;
-    text-decoration-line: underline;
-  }
-`;
-
-//NavLogo
-export const NavLogo = styled.div`
-  width: 113.33px;
-  height: 43px;
-  padding: 13px 0 13px 0;
-  background-color: transparent;
-  margin-bottom: 20px;
-`;
-
-export const NavImg = styled.img`
-  width: 113.33px;
-  height: 17px;
-  color: #181818;
-`;
-
-//NavMenu
-export const MenuNav = styled.nav`
-  display: block;
-  visibility: visible;
-`;
-
-export const MenuList = styled.ul`
-  padding: 18px 0 10px 0;
-`;
-
-//NavMenuItem
-export const MenuItem = styled.li`
-  padding: 5px 0;
-  margin-bottom: 16px;
-`;
-
-export const MenuLink = styled.a`
-  color: #ffffff;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-`;
-
-//PlaylistItem
-export const PlaylistItem = styled.div`
-  width: 100%;
-  display: block;
-  margin-bottom: 12px;
-`;
-
-export const PlaylistTrack = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-`;
-
-export const TrackTitle = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  width: 447px;
-`;
-
-export const TrackTitleImage = styled.div`
-  width: 51px;
-  height: 51px;
-  padding: 16px;
-  background: #313131;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
-  margin-right: 17px;
-`;
-
-export const TrackTitleSvg = styled.svg`
-  width: 18px;
-  height: 17px;
-  fill: transparent;
-  stroke: #4e4e4e;
-`;
-
-export const TrackTitleLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-`;
-
-export const TrackTitleSpan = styled.span`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #4e4e4e;
-`;
-
-export const TrackAuthor = styled.div`
-  width: 321px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: start;
-  -ms-flex-pack: start;
-  justify-content: flex-start;
-`;
-
-export const TrackAuthorLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-  text-align: left;
-`;
-
-export const TrackAlbum = styled.div`
-  width: 245px;
-`;
-
-export const TrackAlbumLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #696969;
-`;
-
-export const TrackTimeSvg = styled.svg`
-  width: 14px;
-  height: 12px;
-  margin-right: 17px;
-  fill: transparent;
-  stroke: #696969;
-`;
-
-export const TrackTimeText = styled.span`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  text-align: right;
-  color: #696969;
-`;
-
-//Search
-export const CenterblockSearch = styled.div`
-  width: 100%;
-  border-bottom: 1px solid #4e4e4e;
-  margin-bottom: 51px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-`;
-
-export const SearchSvg = styled.svg`
-  width: 17px;
-  height: 17px;
-  margin-right: 5px;
-  stroke: #ffffff;
-  fill: transparent;
-`;
-
-export const SearchText = styled.input`
-  -webkit-box-flex: 100;
-  -ms-flex-positive: 100;
-  flex-grow: 100;
-  background-color: transparent;
-  border: none;
-  padding: 13px 10px 14px;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-  &::-webkit-input-placeholder {
-    background-color: transparent;
-    color: #ffffff;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-  }
-  &:-ms-input-placeholder {
-    background-color: transparent;
-    color: #ffffff;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-  }
-  &::-ms-input-placeholder {
-    background-color: transparent;
-    color: #ffffff;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-  }
-  &::placeholder {
-    background-color: transparent;
-    color: #ffffff;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-  }
-`;
-
-//TrackPlay
-export const PlayerTrackPlay = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-`;
-
-export const TrackPlayContain = styled.div`
-  width: auto;
-  display: -ms-grid;
-  display: grid;
-  -ms-grid-columns: auto 1fr;
-  grid-template-columns: auto 1fr;
-  grid-template-areas: "image author" "image album";
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-`;
-
-export const TrackPlayImage = styled.div`
-  width: 51px;
-  height: 51px;
-  background-color: #313131;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
-  margin-right: 12px;
-  -ms-grid-row: 1;
-  -ms-grid-row-span: 2;
-  -ms-grid-column: 1;
-  grid-area: image;
-`;
-
-export const TrackPlaySvg = styled.svg`
-  width: 18px;
-  height: 17px;
-  fill: transparent;
-  stroke: #4e4e4e;
-`;
-
-export const TrackPlayAuthor = styled.div`
-  -ms-grid-row: 1;
-  -ms-grid-column: 2;
-  grid-area: author;
-  min-width: 49px;
-`;
-
-export const TrackPlayAuthorLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-  white-space: nowrap;
-`;
-
-export const TrackPlayAlbum = styled.div`
-  -ms-grid-row: 2;
-  -ms-grid-column: 2;
-  grid-area: album;
-  min-width: 49px;
-`;
-
-export const TrackPlayAlbumLink = styled.a`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 13px;
-  line-height: 24px;
-  color: #ffffff;
-`;
-
-export const TrackPlayLikeDis = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-left: 26%;
-`;
-
-export const TrackPlayLike = styled.div`
-  padding: 5px;
-  &:active {
-    border-color: #ad61ff;
-    color: #ad61ff;
-    cursor: pointer;
-  }
-`;
-
-export const TrackPlayLikeSvg = styled.svg`
-  width: 14px;
-  height: 12px;
-  fill: transparent;
-  stroke: #696969;
-  &:active {
-    fill: transparent;
-    stroke: #ffffff;
-    cursor: pointer;
-  }
-  &:hover {
-    fill: transparent;
-    stroke: #acacac;
-    cursor: pointer;
-  }
-`;
-
-export const TrackPlayDislike = styled.div`
-  padding: 5px;
-  margin-left: 28.5px;
-  &:active {
-    border-color: #ad61ff;
-    color: #ad61ff;
-    cursor: pointer;
-  }
-`;
-
-export const TrackPlayDislikeSvg = styled.svg`
-  width: 14.34px;
-  height: 13px;
-  fill: transparent;
-  stroke: #696969;
-  &:active {
-    fill: transparent;
-    stroke: #ffffff;
-    cursor: pointer;
-  }
-  &:hover {
-    fill: transparent;
-    stroke: #acacac;
-    cursor: pointer;
-  }
-`;
-
-//App
-export const Wrapper = styled.div`
-  width: 100%;
-  min-height: 100%;
-  overflow: hidden;
-`;
-
-export const Container = styled.div`
-  max-width: 1920px;
-  height: 100vh;
-  margin: 0 auto;
-  position: relative;
-  background-color: #181818;
-`;
-
-export const Main = styled.main`
-  -webkit-box-flex: 1;
-  -ms-flex: 1 1 auto;
-  flex: 1 1 auto;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-  flex-wrap: wrap;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-`;
-
-export const MainNav = styled.nav`
-  width: 244px;
-  background-color: #181818;
-  padding: 20px 0 20px 36px;
-`;
-
-export const MainCenterblock = styled.div`
-  width: auto;
-  -webkit-box-flex: 3;
-  -ms-flex-positive: 3;
-  flex-grow: 3;
-  padding: 20px 40px 20px 111px;
-`;
-
-export const CenterblockH2 = styled.h2`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 64px;
-  line-height: 72px;
-  letter-spacing: -0.8px;
-  margin-bottom: 45px;
-`;
-
-export const CenterblockContent = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-`;
-
-export const ContentPlaylist = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-  overflow-y: auto;
-`;
-
-export const MainSidebar = styled.div`
-  max-width: 418px;
-  padding: 20px 90px 20px 78px;
-`;
-
-export const SidebarPersonal = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-pack: end;
-  -ms-flex-pack: end;
-  justify-content: flex-end;
-  padding: 12px 0 15px 0;
-`;
-
-export const SidebarPersonalName = styled.p`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-  margin-right: 16px;
-`;
-
-export const SidebarIcon = styled.div`
-  width: 43px;
-  height: 43px;
-  background-color: #313131;
-  border-radius: 50%;
-  cursor: pointer;
-`;
-
-export const SidebarBlock = styled.div`
-  height: 100%;
-  padding: 240px 0 0 0;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-  -webkit-box-pack: start;
-  -ms-flex-pack: start;
-  justify-content: flex-start;
-`;
-
-export const SidebarList = styled.div`
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-`;
-
-export const SidebarItem = styled.div`
-  width: 250px;
-  height: 150px;
-  &:not(:last-child) {
-    margin-bottom: 30px;
-  }
-`;
-
-export const SidebarLink = styled.a`
-  width: 100%;
-  height: 100%;
-`;
-
-export const SidebarImg = styled.img`
-  width: 100%;
-  height: auto;
-`;
