+++ src/components/Wrapper/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  width: 100%;
+  min-height: 100%;
+  overflow: hidden;
+`;
